Make newXMLHttpRequest async, handles errors (#34)
The newXMLHttpRequest presented two issues:

1. Creating an XHR connection in at least one browser (Chrome) blocks
while creating the connection, despite the result coming back
asynchonrously. This creates a certain lag when calling the function,
and thus also causes event handlers to block which delays subsequent
event handlers from running. A simple example is an event containing a
request; there are two handlers for this event (1) a spinner, (2) one
that makes a connection. The one that makes a connection will stop the
spinner from activating. If there is sufficient network delay, the
spinner won't appear for some 300ms to some seconds, defeating the
purpose of a spinner in the first place.

2. Handling an error. Related to the first, creating the XHR connection
blocks but can then throw an XHRError exception, which was not handled
anywhere in the code. Connection timeout is a common issue that
we'd like to handle in the UI (think GMail or Slack displaying a
"disconnected; reconnecting in n seconds..." kind of message), so I've
put the exception in the event's output so that a user-friendly UI
message can be displayed.

I've added some documentation for related functions.